backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.4
requests==2.31.0
google-genai==0.1.2
resend==0.8.0
python-dotenv==1.0.1
